Added "DoWith" function and fleshed out object test.
Added "DoWith" function to objects and fleshed out object test.

The DoWith function allows you to execute code on within an object.
You can pass it "all" to execute code on every object one-by-one, an id to execute code on nothing but that specific object, or a table of ids to execute code on each id in the table.

If you give DoWith an id which doesn't exist or an invalid id, it'll simply ignore running code on the id(s) in question, and move on to the next id (If any.) without trouble.