Lots of changes! (Happy Easter, btw! :) :candy: )
Lots of engine changes!
The biggest one being *drumroll please*...

**ANIMATIONS!!!** :smiley:
Sprites FINALLY animate! (It's about time!) And it's easy as pie, too!
An example for this is coming soon, but in the mean-time, here's how you do some simple sprite-animation:

To load a multi-frame sprite, simply do "Sprites.Add({"frame1.png","frame2.png",etc.})", replacing the values in quotes (And the "etc.", of course.) with the names of your frames.

To change the speed of the animation (Which is in how many frames it takes to go to the next frame, **NOT** the FPS! The bigger the number the slower it gets, not the other way around.), simply do "object.animspeed = 5", replacing "object" with the name of your object and "5" with your animation speed.

And that's it! Sprites auto-matically animate, so you need do nothing else! :)

Other features include:

New, fully-working frame-rate variable which makes setting the frame-rate a **LOT** easier!
For example, to set the frame-rate to 30 fps, just do "framerate = 30"! Done!

New object:changesprite function to make changing sprites a breeze!
You can still set object.sprite to a different sprite instead if you wish, but this comes at the risk of crashing the game unless you set object.currentframe to 0 as well. object.changesprite does this for you.

The Objects.Add function now takes an extra parameter: The speed of the sprite's animation. (Same as setting object.animspeed as described above.)